Refactor TorrentProvider.py -> torrent_provider.py
