+++ medusa/providers/torrent/html/animebytes.py
@@ -24,7 +24,7 @@ import traceback
 from requests.compat import urljoin
 from requests.utils import dict_from_cookiejar
 from six.moves.urllib_parse import parse_qs
-from ..TorrentProvider import TorrentProvider
+from ..torrent_provider import TorrentProvider
 from .... import logger, tv_cache
 from ....bs4_parser import BS4Parser
 from ....helper.common import convert_size
