+++ medusa/providers/torrent/html/cpasbien.py
@@ -21,7 +21,7 @@ from __future__ import unicode_literals
 import re
 import traceback
 
-from ..TorrentProvider import TorrentProvider
+from ..torrent_provider import TorrentProvider
 from .... import logger, tv_cache
 from ....bs4_parser import BS4Parser
 from ....helper.common import convert_size, try_int
