+++ medusa/providers/torrent/html/extratorrent.py
@@ -23,7 +23,7 @@ from requests.compat import urljoin
 
 import validators
 
-from ..TorrentProvider import TorrentProvider
+from ..torrent_provider import TorrentProvider
 from .... import logger, tv_cache
 from ....bs4_parser import BS4Parser
 from ....helper.common import convert_size, try_int
