+++ medusa/providers/torrent/html/hdspace.py
@@ -26,7 +26,7 @@ import traceback
 from requests.compat import urljoin
 from requests.utils import dict_from_cookiejar
 
-from ..TorrentProvider import TorrentProvider
+from ..torrent_provider import TorrentProvider
 from .... import logger, tv_cache
 from ....bs4_parser import BS4Parser
 from ....helper.common import convert_size, try_int
