+++ medusa/providers/torrent/html/limetorrents.py
@@ -24,7 +24,7 @@ import traceback
 from contextlib2 import suppress
 import requests
 from requests.compat import urljoin
-from ..TorrentProvider import TorrentProvider
+from ..torrent_provider import TorrentProvider
 from .... import logger, tv_cache
 from ....bs4_parser import BS4Parser
 from ....helper.common import convert_size, try_int
