+++ medusa/providers/torrent/html/xthor.py
@@ -23,7 +23,7 @@ import re
 import traceback
 
 from requests.utils import dict_from_cookiejar
-from ..TorrentProvider import TorrentProvider
+from ..torrent_provider import TorrentProvider
 from .... import logger, tv_cache
 from ....bs4_parser import BS4Parser
 from ....helper.common import convert_size, try_int
