+++ medusa/providers/torrent/json/bitcannon.py
@@ -22,7 +22,7 @@ import traceback
 
 from requests.compat import urljoin
 import validators
-from ..TorrentProvider import TorrentProvider
+from ..torrent_provider import TorrentProvider
 from .... import logger, tv_cache
 from ....helper.common import convert_size, try_int
