+++ medusa/providers/torrent/json/btdigg.py
@@ -23,7 +23,7 @@ import traceback
 
 from requests.compat import urljoin
 
-from ..TorrentProvider import TorrentProvider
+from ..torrent_provider import TorrentProvider
 from .... import logger, tv_cache
 from ....helper.common import convert_size
