+++ medusa/providers/torrent/json/hdbits.py
@@ -20,7 +20,7 @@ from __future__ import unicode_literals
 import json
 
 from requests.compat import urlencode, urljoin
-from ..TorrentProvider import TorrentProvider
+from ..torrent_provider import TorrentProvider
 from .... import logger, tv_cache
 from ....helper.exceptions import AuthException
