+++ medusa/providers/torrent/xml/bitsnoop.py
@@ -22,7 +22,7 @@ import traceback
 
 import medusa as app
 from requests.compat import urljoin
-from ..TorrentProvider import TorrentProvider
+from ..torrent_provider import TorrentProvider
 from .... import logger, tv_cache
 from ....bs4_parser import BS4Parser
 from ....helper.common import convert_size, try_int
