+++ pytest.ini
@@ -95,7 +95,7 @@ flake8-ignore =
     medusa/providers/nzb/newznab.py D100 D102 D200 D204 D205 D400 D401 N803 W391
     medusa/providers/nzb/omgwtfnzbs.py D100 D101 D102 N802 N803
     medusa/providers/nzb/womble.py D100 D101 D102 D204 D400 N802
-    medusa/providers/torrent/TorrentProvider.py D100 D101 D102 D205 D400
+    medusa/providers/torrent/torrent_provider.py D100 D101 D102 D205 D400
     medusa/providers/torrent/__init__.py D104
     medusa/providers/torrent/html/__init__.py D104
     medusa/providers/torrent/html/abnormal.py D100 D102 D204 D400
