+++ tests/legacy/providers/torrent_provider_tests.py
@@ -22,7 +22,7 @@ import os
 
 import medusa as app
 from medusa.providers.generic_provider import GenericProvider
-from medusa.providers.torrent.TorrentProvider import TorrentProvider
+from medusa.providers.torrent.torrent_provider import TorrentProvider
 from six import iteritems
 from .generic_provider_tests import GenericProviderTests
